requirements.txt
SpeechRecognition==3.10.0
openai==1.12.0
python-dotenv==1.0.0
soundfile==0.12.1
numpy==1.24.3
scipy==1.11.2
TTS==0.21.1
pymongo==4.6.1
httpx==0.26.0
